Refactor: widgets 레이어에서  comment,post 바로 하위에 ui컴포넌트 배치
